requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic-settings==2.2.1
jinja2==3.1.4
requests==2.31.0
google-generativeai==0.5.4
feedparser==6.0.11
beautifulsoup4==4.12.3
sumy==0.11.0
nltk==3.8.1
premailer==3.10.0
pytest==8.2.1
pytest-mock==3.12.0
numpy==2.3.2
tweepy==4.16
psycopg2-binary==2.9.10